Add new channel-based architecture (#12)
Now each layer is a tokio task spawned.
* first is the server that listens to udp socket and forwards to the
transport layer. It also listens for messages from transport layer and
forwards those to udp
* then it's the transport layer that listens from the server and
forwards to core. It also listens from core and forwards to server.
* then it's the core that basically handles the sip messages (atm it
handles the registration)

+ cargo fmt
+ cargo clippy

We moved the processor inside the core layer.

there are some things that need to be done:
* remove transactions/dialogs traits/models and do them differently
* add tests per layer, should be easy but we need to implement some
infra first for tests
* implement the layers based on traits